[modify] support 64bit large file; a little slow on 32-bit

+++ BufferedFileBinaryInputOutputPort.h
@@ -48,7 +48,7 @@ public:
     Object position() const;
     int close();
     int pseudoClose();
-    bool setPosition(int position) ;
+    bool setPosition(int64_t position);
     ucs4string toString();
 
     // binary port interfaces
@@ -58,15 +58,15 @@ public:
     // input interfaces
     int getU8();
     int lookaheadU8();
-    int readBytes(uint8_t* buf, int reqSize, bool& isErrorOccured);
-    int readSome(uint8_t** buf, bool& isErrorOccured);
-    int readAll(uint8_t** buf, bool& isErrorOccured);
+    int64_t readBytes(uint8_t* buf, int64_t reqSize, bool& isErrorOccured);
+    int64_t readSome(uint8_t** buf, bool& isErrorOccured);
+    int64_t readAll(uint8_t** buf, bool& isErrorOccured);
 
     // output interfaces
     int putU8(uint8_t v);
-    int putU8(uint8_t* v, int size);
-    int putByteVector(ByteVector* bv, int start = 0);
-    int putByteVector(ByteVector* bv, int start, int count);
+    int64_t putU8(uint8_t* v, int64_t size);
+    int64_t putByteVector(ByteVector* bv, int64_t start = 0);
+    int64_t putByteVector(ByteVector* bv, int64_t start, int64_t count);
     void flush();
     void internalFlush();
     File* getFile();
@@ -76,10 +76,10 @@ protected:
         BUF_SIZE = 8192,
     };
 
-    void forwardPosition(int offset);
+    void forwardPosition(int64_t offset);
     void initializeBuffer();
-    virtual int writeToBuffer(uint8_t* buf, size_t size) = 0;
-    int readFromBuffer(uint8_t* dest, int reqSize);
+    virtual int64_t writeToBuffer(uint8_t* buf, int64_t size) = 0;
+    int64_t readFromBuffer(uint8_t* dest, int64_t reqSize);
     void fillBuffer();
     bool isBufferDirty() { return isDirty_; }
     void invalidateBuffer();
@@ -88,11 +88,11 @@ protected:
     ucs4string fileName_;
     uint8_t* buffer_;
     bool isDirty_;
-    int position_;
+    int64_t position_;
     bool isClosed_;
     bool isPseudoClosed_;
-    int bufferSize_;
-    int bufferIndex_;
+    int64_t bufferSize_;
+    int64_t bufferIndex_;
 };
 
 } // namespace scheme
