+++ Codec.cpp
@@ -1,5 +1,5 @@
 /*
- * Codec.cpp - 
+ * Codec.cpp -
  *
  *   Copyright (c) 2008  Higepon(Taro Minowa)  <higepon@users.sourceforge.jp>
  *
@@ -39,5 +39,5 @@ int Codec::putChar(BinaryOutputPort* port, ucs4char c, enum ErrorHandlingMode mo
 {
     static uint8_t buf[4];
     const int size = putChar(buf, c, mode);
-    return port->putU8(buf, size);
+    return static_cast<int>(port->putU8(buf, size));
 }
