+++ FileBinaryInputPort.h
@@ -51,9 +51,9 @@ public:
     ucs4string toString();
     int getU8();
     int lookaheadU8();
-    int readBytes(uint8_t* buf, int reqSize, bool& isErrorOccured);
-    int readSome(uint8_t** buf, bool& isErrorOccured);
-    int readAll(uint8_t** buf, bool& isErrorOccured);
+    int64_t readBytes(uint8_t* buf, int64_t reqSize, bool& isErrorOccured);
+    int64_t readSome(uint8_t** buf, bool& isErrorOccured);
+    int64_t readAll(uint8_t** buf, bool& isErrorOccured);
     int open();
     int close();
     int pseudoClose();
@@ -61,7 +61,7 @@ public:
     bool hasPosition() const;
     bool hasSetPosition() const;
     Object position() const;
-    bool setPosition(int position);
+    bool setPosition(int64_t position);
     File* getFile() { return file_; }
 
 private:
@@ -72,7 +72,7 @@ private:
     bool isClosed_;
     bool isPseudoClosed_;
     int aheadU8_;
-    int position_;
+    int64_t position_;
 };
 
 } // namespace scheme
