+++ FileBinaryOutputPort.h
@@ -47,9 +47,9 @@ public:
     virtual ~FileBinaryOutputPort();
 
     int putU8(uint8_t v);
-    int putU8(uint8_t* v, int size);
-    int putByteVector(ByteVector* bv, int start = 0);
-    int putByteVector(ByteVector* bv, int start, int count);
+    int64_t putU8(uint8_t* v, int64_t size);
+    int64_t putByteVector(ByteVector* bv, int64_t start = 0);
+    int64_t putByteVector(ByteVector* bv, int64_t start, int64_t count);
     int open();
     int close();
     int pseudoClose();
@@ -59,7 +59,7 @@ public:
     bool hasPosition() const;
     bool hasSetPosition() const;
     Object position() const;
-    bool setPosition(int position);
+    bool setPosition(int64_t position);
     File* getFile();
 
 protected:
@@ -67,7 +67,7 @@ protected:
     ucs4string fileName_;
     bool isClosed_;
     bool isPseudoClosed_;
-    int position_;
+    int64_t position_;
 };
 
 } // namespace scheme
