+++ Fixnum.h
@@ -47,7 +47,7 @@ public:
         MIN = -MAX - 1
     };
 
-    static bool canFit(long long n)
+    static bool canFit(int64_t n)
     {
         return Fixnum::MIN <= n && n <= Fixnum::MAX;
     }
