+++ StandardInputPort.h
@@ -57,7 +57,7 @@ public:
     Object position() const {
         return Object::Undef;
     }
-    bool setPosition(int position)
+    bool setPosition(int64_t position)
     {
         return false;
     }
